extract funtions and components to uitls file, add descriptive comments
